+++ src/apps/testapps/testCellToLocalIj.c
@@ -263,11 +263,11 @@ SUITE(h3ToLocalIj) {
 
         for (uint32_t i = 1; i <= 32; i++) {
             CoordIJ ij2;
-            t_assert(H3_EXPORT(cellToLocalIj)(cell, cell, i, &ij2) == E_FAILED,
+            t_assert(H3_EXPORT(cellToLocalIj)(cell, cell, i, &ij2) == E_OPTION_INVALID,
                      "Invalid mode fail for cellToLocalIj");
             H3Index cell2;
             t_assert(
-                H3_EXPORT(localIjToCell)(cell, &ij2, i, &cell2) == E_FAILED,
+                H3_EXPORT(localIjToCell)(cell, &ij2, i, &cell2) == E_OPTION_INVALID,
                 "Invalid mode fail for cellToLocalIj");
         }
     }