+++ src/apps/testapps/testPolygonToCells.c
@@ -400,8 +400,8 @@ SUITE(polygonToCells) {
         int64_t numHexagons;
         for (uint32_t flags = 1; flags <= 32; flags++) {
             t_assert(
-                H3_EXPORT(maxPolygonToCellsSize)(&sfGeoPolygon, 9, flags,
-                                                 &numHexagons) == E_FAILED,
+                H3_EXPORT(maxPolygonToCellsSize)(
+                    &sfGeoPolygon, 9, flags, &numHexagons) == E_OPTION_INVALID,
                 "Flags other than 0 are invalid for maxPolygonToCellsSize");
         }
         t_assertSuccess(H3_EXPORT(maxPolygonToCellsSize)(&sfGeoPolygon, 9, 0,
@@ -409,7 +409,7 @@ SUITE(polygonToCells) {
         H3Index *hexagons = calloc(numHexagons, sizeof(H3Index));
         for (uint32_t flags = 1; flags <= 32; flags++) {
             t_assert(H3_EXPORT(polygonToCells)(&sfGeoPolygon, 9, flags,
-                                               hexagons) == E_FAILED,
+                                               hexagons) == E_OPTION_INVALID,
                      "Flags other than 0 are invalid for polygonToCells");
         }
         free(hexagons);