+++ src/h3lib/lib/algos.c
@@ -751,7 +751,7 @@ H3Error H3_EXPORT(gridRingUnsafe)(H3Index origin, int k, H3Index *out) {
 H3Error H3_EXPORT(maxPolygonToCellsSize)(const GeoPolygon *geoPolygon, int res,
                                          uint32_t flags, int64_t *out) {
     if (flags != 0) {
-        return E_FAILED;
+        return E_OPTION_INVALID;
     }
     // Get the bounding box for the GeoJSON-like struct
     BBox bbox;
@@ -858,7 +858,7 @@ H3Error _getEdgeHexagons(const GeoLoop *geoloop, int64_t numHexagons, int res,
 H3Error H3_EXPORT(polygonToCells)(const GeoPolygon *geoPolygon, int res,
                                   uint32_t flags, H3Index *out) {
     if (flags != 0) {
-        return E_FAILED;
+        return E_OPTION_INVALID;
     }
     // One of the goals of the polygonToCells algorithm is that two adjacent
     // polygons with zero overlap have zero overlapping hexagons. That the